+++ src/components/Pagination/Pagination.tsx
@@ -1,32 +1,18 @@
 import { useCallback } from 'react';
-import { Button, ButtonGroup, Flex } from '@chakra-ui/react';
+import { Box, Button, ButtonGroup, Flex } from '@chakra-ui/react';
 
 import { IPagination } from '@interfaces/pagination.interfaces';
 import { getVisiblePages } from '@utils/pagination.utils';
 
 const Pagination = <Data,>(props: IPagination<Data>) => {
-  const { data, setPaginationData, pageIndex, pageSize } = props;
-
-  const countPages = localStorage.getItem('PAGE_PASS_COUNT');
-  // totalPages - кол-во всех страниц данного типа (самолетов и тд.), если нет значения в localStorage, высчитываем ее самостоятельно (второй вариант только для самолетов)
-  const totalPages =
-    countPages !== null
-      ? JSON.parse(countPages)
-      : data
-      ? data.length / pageSize
-      : 1;
+  const { data, setPaginationData, pageIndex, pageSize, totalPages = 1 } = props;
 
   const setPagination = useCallback(
     (pageNumber: number) => {
-      if (data?.length) {
-        const airplanesLength = data.length;
-        if (pageNumber >= 0 && pageNumber < airplanesLength / pageSize) {
-          setPaginationData(pageNumber);
-        }
+      if (data?.length && pageNumber >= 0) {
+        setPaginationData(pageNumber);
       }
-    },
-    [data?.length, pageSize]
-  );
+    }, [data?.length, setPaginationData]);
 
   const handleToFirstPage = (currentPage: number): void => {
     if (currentPage !== 0) {
@@ -46,9 +32,10 @@ const Pagination = <Data,>(props: IPagination<Data>) => {
     }
   };
 
-  return data ? (
+  return data && (totalPages>1) ? (
     <Flex my={8}>
       <Button
+        display={pageIndex == 0 ? 'none' : 'block'}
         me={2}
         onClick={() => {
           handleToFirstPage(pageIndex);
@@ -71,6 +58,7 @@ const Pagination = <Data,>(props: IPagination<Data>) => {
         {'<<'}
       </Button>
       <Button
+        display={pageIndex == 0 ? 'none' : 'block'}
         me={5}
         className="border rounded p-1"
         onClick={() => setPagination(pageIndex - 1)}
@@ -98,9 +86,7 @@ const Pagination = <Data,>(props: IPagination<Data>) => {
             onClick={() => handleToPressedPage(pageIndex, page)}
             borderRadius="0.125rem"
             border="0.0625rem solid #DEDEDE"
-            bgColor={
-              page === pageIndex + 1 ? '#398AEA' : 'rgba(217, 217, 217, 0.15)'
-            }
+            bgColor={page === pageIndex + 1 ? '#398AEA' : 'rgba(217, 217, 217, 0.15)'}
             color={page === pageIndex + 1 ? '#FFFFFF' : '#393939'}
             _hover={{
               backgroundColor: '#398AEA',
@@ -116,8 +102,37 @@ const Pagination = <Data,>(props: IPagination<Data>) => {
             {page}
           </Button>
         ))}
+        <Box 
+          pe={2}
+          ps={2}
+          pt={2}
+          display={ totalPages > 5 && pageIndex < totalPages - 2 ? 'block':'none'}>
+            ...
+        </Box>
+        <Button
+        display={totalPages > 5 && pageIndex < totalPages - 2 ? 'block' : 'none'}
+        className="border rounded p-1"
+        onClick={() => setPagination(totalPages - 1)}
+        borderRadius="0.125rem"
+        border="0.0625rem solid #DEDEDE"
+        bgColor="rgba(217, 217, 217, 0.15)"
+        color="#393939"
+        _hover={{
+          backgroundColor: '#398AEA',
+          borderColor: '#398AEA',
+          color: '#ffffff',
+        }}
+        _active={{
+          backgroundColor: '#398AEA',
+          borderColor: '#398AEA',
+          color: '#ffffff',
+        }}
+      >
+        {totalPages}
+      </Button>
       </ButtonGroup>
       <Button
+        display={pageIndex + 1 == totalPages ? 'none' : 'block'}
         ms={5}
         className="border rounded p-1"
         onClick={() => setPagination(pageIndex + 1)}
@@ -139,12 +154,10 @@ const Pagination = <Data,>(props: IPagination<Data>) => {
         {'>'}
       </Button>
       <Button
+        display={pageIndex + 1 == totalPages ? 'none' : 'block'}
         ms={2}
         className="border rounded p-1"
-        onClick={() => {
-          const pageCount = Math.ceil(data.length / pageSize - 1);
-          handleToLastPage(pageIndex, pageCount);
-        }}
+        onClick={() => setPagination(totalPages-1)}
         borderRadius="0.125rem"
         border="0.0625rem solid #DEDEDE"
         bgColor="rgba(217, 217, 217, 0.15)"